# internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Simple Task Manager API
  version: 1.5.0
  description: |
    REST API for user authentication and per-user task management.
    Protected task endpoints require a Bearer JWT access token.
    This document is the governed public compatibility contract for API v1.
...
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/workspaces:
    get:
      tags: [Workspaces]
      summary: List workspaces visible to the current user
      operationId: listWorkspaces
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Workspace memberships
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAccessListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Workspaces]
      summary: Create a shared workspace
      operationId: createWorkspace
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created with current user as owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAccessEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          example: Platform Team
...
    Workspace:
      type: object
      required: [id, name, created_by_user_id, is_personal, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        created_by_user_id:
          type: integer
          format: int64
        is_personal:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkspaceAccess:
      allOf:
        - $ref: '#/components/schemas/Workspace'
        - type: object
          required: [role]
          properties:
            role:
              type: string
              enum: [owner, admin, member]
...
    WorkspaceAccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/WorkspaceAccess'
    WorkspaceAccessListEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceAccess'
...
    ErrorEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
...